Agregar clase Departmento con atributos y métodos para gestión de departamentos y empleados.
